# docs/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Doctor 1 — Clinic Booking & Queue API
  version: "1.0.0"
  description: |
    The complete v1 API, consumed by the Flutter app.

    **This file is the source of truth.** The Postman collection is generated
    from it (`php artisan api:postman`), and a test fails if a registered route
    is missing here — so it cannot drift from the code.

    ## The envelope

    Every response is one of three shapes, including 404s and 500s:

    ```
    { "status": "success", "message": "...", "data": { } }
    { "status": "error",   "message": "...", "error": { code, details, fields } }
    ```

    Branch on `error.code` — a stable UPPER_SNAKE string — never on `message`,
    which is localised prose meant to be shown to the user as-is.

    ## value + display

    Anything rendered as text arrives pre-formatted, driven by
    `Accept-Language`, so the client never re-implements Arabic formatting.
    Use `value` for logic, show `display`.

    ## Clinic scope

    The clinic comes from the token. **No endpoint accepts `clinic_id`.**
    Another clinic's record returns 404, never its data.
  contact:
    name: Backend
    url: https://github.com/
servers:
  - url: "{base_url}/api/v1"
    variables:
      base_url:
        default: http://127.0.0.1:8000
        description: Host without a trailing slash
security:
  - bearerAuth: []
...
paths:
...
  /postpone:
    post:
      tags: [Postpone]
      summary: Postpone today
      description: |
        Cancels the selected bookings with reason `emergency`, **frees their
        slots**, and returns the call list.

        WhatsApp template broadcasts are handled separately through
        `/broadcasts`. Omit `booking_ids` entirely for "كل المريضات".
      operationId: postpone
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                booking_ids:
                  type: array
                  items: { type: integer }
                  description: Omit for everyone still pending.
                  example: [88, 91]
      responses:
        "200":
          description: Postponed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: "#/components/schemas/CallListRow" }
                          postponed_count: { type: integer }
        "400":
          description: "`NOTHING_TO_POSTPONE` — including when the only patient left is with the doctor"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The Sanctum token from `POST /auth/login`.
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, message, data]
      properties:
        status: { type: string, enum: [success] }
        message: { type: string, description: Localised; safe to show as-is. }
    ErrorEnvelope:
      type: object
      required: [status, message, error]
      properties:
        status: { type: string, enum: [error] }
        message: { type: string }
        error:
          type: object
          required: [code, details, fields]
          properties:
            code:
              type: string
              description: Stable UPPER_SNAKE. **Branch on this, never on `message`.**
            details:
              type: [object, "null"]
              description: Structured context for a few errors.
            fields:
              type: [object, "null"]
              description: Present on 422 only — field name to messages.
              additionalProperties:
                type: array
                items: { type: string }
    Display:
      type: object
      description: Machine value plus a ready-to-render string.
      required: [value, display]
      properties:
        value: { type: [string, integer] }
        display: { type: string }
...
    CallListRow:
      type: object
      properties:
        booking_id: { type: integer }
        patient:
          type: [object, "null"]
          properties:
            id: { type: integer }
            code: { type: string }
            name: { type: string }
            phone:
              allOf: [{ $ref: "#/components/schemas/Display" }]
              description: Unmasked — every row has a call action.
        visit_type:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
        original_date: { $ref: "#/components/schemas/Display" }
        original_start_time: { $ref: "#/components/schemas/Display" }
        contacted: { type: boolean }
        contacted_at: { type: [string, "null"], format: date-time }
...